+++ src/plonemeeting/portal/core/migrations/migrate_to_1003.py
@@ -1,42 +1,14 @@
 # -*- coding: utf-8 -*-
 
+import logging
+
 from imio.helpers.content import richtextval
 from imio.migrator.migrator import Migrator
-from plone import api
-from plonemeeting.portal.core.config import CONFIG_FOLDER_ID
-from plonemeeting.portal.core.config import FACETED_FOLDER_ID
-from plonemeeting.portal.core.config import FACETED_XML_PATH
-
-import logging
-import os
 
 logger = logging.getLogger("plonemeeting.portal.core")
 
 
 class MigrateTo1003(Migrator):
-
-    def _merge_faceted(self):
-        logger.info("Merging faceted folders for every institutions...")
-        institutions = [obj for obj in self.portal.objectValues()
-                        if obj.portal_type == "Institution"]
-        # remove "decisions" folder from every institutions
-        for institution in institutions:
-            decisions = [obj for obj in institution.objectValues()
-                         if obj.portal_type == "Folder" and obj.getId() == "decisions"]
-            if decisions:
-                api.content.delete(decisions[0])
-        # re-apply faceted config
-        faceted = self.portal.get(CONFIG_FOLDER_ID).get(FACETED_FOLDER_ID)
-        subtyper = faceted.restrictedTraverse("@@faceted_subtyper")
-        subtyper.enable()
-        # file is one level up, we are in migrations folder
-        faceted_config_path = os.path.join(os.path.dirname(__file__), "..", FACETED_XML_PATH)
-        with open(faceted_config_path, "rb") as faceted_config:
-            faceted.unrestrictedTraverse("@@faceted_exportimport").import_xml(
-                import_file=faceted_config
-            )
-        logger.info("Done.")
-
     def _fix_formatted_title(self):
         """
         Fix formatted_title on existing items where formatted_title is null
@@ -47,14 +19,13 @@ class MigrateTo1003(Migrator):
             item = brain.getObject()
             if not item.formatted_title:
                 item.formatted_title = richtextval("<p>" + item.title + "</p>")
-        logger.info("Reindexing SearchableText")
-        self.reindexIndexes(idxs=["SearchableText"], update_metadata=True)
         logger.info("Done.")
 
     def run(self):
         logger.info("Migrating to plonemeeting.portal 1003...")
+
         self._fix_formatted_title()
-        self._merge_faceted()
+        self.reindexIndexes(idxs=["SearchableText", "pretty_representatives"], update_metadata=True)
 
 
 def migrate(context):
@@ -62,7 +33,8 @@ def migrate(context):
     This migration function will:
 
        1) Update the registry to add new bundles;
-       2) Merge faceted folders.
+       2) Reindex indexes "SearchableText" (related to "_fix_formatted_title" step)
+          and "pretty_representatives" now that indexed order was fixed.
     """
     migrator = MigrateTo1003(context)
     migrator.run()